Set the index URL to point to the API doc
This patch changes the swagger UI url to the API doc instead of the
default example.

The only change that was made was changing the url in src/main/html/index.html
and calling npm run build.

After this patch when pointing the browswer to http://localhost:10000/ui
you will get the swager ui with scylla's api doc.

Signed-off-by: Amnon Heiman <amnon@scylladb.com>